program: add abstract for my synthesizer-llvm tutorial
A referee wanted to see the link targets explicitly.
I do not know whether he thought of the HTML presentation
when asking for it or whether he refered to a potential printed program.
At least on the website it would look ugly.
Thus I switched back to simple link descriptions.